pyproject.toml
[tool.poetry]
name = "async-nookipedia"
version = "0.1.3"
description = "async API wrapper for Nookipedia API written in python"
authors = ["makubob <makupi@icloud.com>"]
license = "GPL-3.0"
packages = [{include = "nookipedia"}]
readme = "README.md"
repository = "https://github.com/makupi/async-nookipedia"

[tool.poetry.dependencies]
python = "^3.6"
aiohttp = "^3.6.2"
importlib_metadata = "^1.6.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
